chore(component,scheduler): rename component
