--- pkg/component/generic/schedule/v0/config/definition.json
@@ -1,19 +1,0 @@
-{
-  "availableTasks": [],
-  "availableEvents": [
-    "EVENT_CRON_JOB_TRIGGERED"
-  ],
-  "documentationUrl": "https://www.instill.tech/docs/component/generic/schedule",
-  "icon": "assets/schedule.svg",
-  "id": "schedule",
-  "public": true,
-  "title": "Schedule",
-  "vendor": "Generic",
-  "description": "Trigger pipelines on a schedule using cron expressions.",
-  "tombstone": false,
-  "type": "COMPONENT_TYPE_GENERIC",
-  "uid": "f2a89c6d-e315-4b7a-9d8f-c83e45a92c10",
-  "version": "0.1.0",
-  "sourceUrl": "https://github.com/instill-ai/pipeline-backend/blob/main/pkg/component/generic/schedule/v0",
-  "releaseStage": "RELEASE_STAGE_ALPHA"
-}
