+++ pkg/component/generic/scheduler/v0/README.mdx
@@ -1,11 +1,11 @@
 ---
-title: "Schedule"
+title: "Scheduler"
 lang: "en-US"
 draft: false
-description: "Learn about how to set up a VDP Schedule component https://github.com/instill-ai/instill-core"
+description: "Learn about how to set up a VDP Scheduler component https://github.com/instill-ai/instill-core"
 ---
 
-The Schedule component is a generic component that allows users to trigger pipelines on a schedule using cron expressions.
+The Scheduler component is a generic component that allows users to trigger pipelines on a scheduler using cron expressions.
 It can carry out the following tasks:
 
 
@@ -18,7 +18,7 @@ It can carry out the following tasks:
 
 ## Configuration
 
-The component definition and tasks are defined in the [definition.json](https://github.com/instill-ai/pipeline-backend/blob/main/pkg/component/generic/schedule/v0/config/definition.json) and [tasks.json](https://github.com/instill-ai/pipeline-backend/blob/main/pkg/component/generic/schedule/v0/config/tasks.json) files respectively.
+The component definition and tasks are defined in the [definition.json](https://github.com/instill-ai/pipeline-backend/blob/main/pkg/component/generic/scheduler/v0/config/definition.json) and [tasks.json](https://github.com/instill-ai/pipeline-backend/blob/main/pkg/component/generic/scheduler/v0/config/tasks.json) files respectively.
