+++ pkg/component/generic/scheduler/v0/config/events.json
