+++ pkg/component/generic/scheduler/v0/event_cron_job_triggered.go
@@ -1,4 +1,4 @@
-package schedule
+package scheduler
 
 import (
 	"context"
