+++ pkg/component/generic/scheduler/v0/main.go
@@ -1,4 +1,4 @@
-package schedule
+package scheduler
 
 //go:generate compogen readme ./config ./README.mdx --extraContents bottom=.compogen/bottom.mdx
